REFACTOR: separate models by orm and odm
