+++ app/callback/track_stream.py
@@ -3,7 +3,7 @@ from tortoise.models import Model
 from tortoise.transactions import in_transaction
 from app.depends.lock import LockLogicDecorator
 from app.models.email_model import EmailTrackingORM
-from app.models.twilio_model import CallTrackingORM, SMSTrackingORM
+from app.models.orm.twilio_model import CallTrackingORM, SMSTrackingORM
 from app.utils.constant import StreamConstant
 from app.utils.transformer import empty_str_to_none
