+++ app/depends/class_dep.py
@@ -11,9 +11,9 @@ from app.interface.email import EmailInterface, EmailSendInterface
 from app.models.call_model import BaseVoiceCallModel
 from app.models.ingest_model import FileUploadDataIngestModel
 from app.models.email_model import CustomEmailModel, EmailStatus, EmailTemplateModel, TrackingEmailEventORM
-from app.models.link_model import LinkORM
+from app.models.orm.link_model import LinkORM
 from app.models.sms_model import OnGoingBaseSMSModel
-from app.models.twilio_model import CallEventORM, CallStatusEnum, SMSEventORM, SMSStatusEnum
+from app.models.orm.twilio_model import CallEventORM, CallStatusEnum, SMSEventORM, SMSStatusEnum
 from app.services.config_service import ConfigService
 from app.container import Get
 from app.utils.constant import ParseStrategy, SpecialKeyAttributesConstant
