+++ app/ressources/link_ressource.py
@@ -23,7 +23,7 @@ from app.services.database.tortoise_service import TortoiseConnectionService
 from app.services.link_service import LinkService
 from fastapi import status
 from fastapi.responses import HTMLResponse, PlainTextResponse, RedirectResponse
-from app.models.link_model import LinkORM,LinkModel, QRCodeModel, UpdateLinkModel
+from app.models.orm.link_model import LinkORM,LinkModel, QRCodeModel, UpdateLinkModel
 from app.utils.constant import  CostConstant, StreamConstant
 from app.utils.helper import  uuid_v1_mc
 from tortoise.transactions import in_transaction
