+++ app/services/mini/webhook/broker_webhook_service.py
@@ -2,7 +2,7 @@ import json
 from typing import Any
 from app.definition._service import DEFAULT_BUILD_STATE, BaseMiniService, MiniService
 from app.interface.webhook_adapter import WebhookAdapterInterface
-from app.models.webhook_model import AuthConfig, KafkaWebhookModel, RedisWebhookModel, SQSWebhookModel
+from app.models.odm.webhook_model import AuthConfig, KafkaWebhookModel, RedisWebhookModel, SQSWebhookModel
 from app.services.config_service import ConfigService
 from app.services.database.redis_service import RedisService
 from app.services.profile_service import ProfileMiniService
