--- app/tools/memory_tool.py
@@ -1,13 +1,0 @@
-from app.definition._tool import ContextPipelineTool, Tool
-from app.services.config_service import ConfigService
-from app.services.database.graphiti_service import GraphitiService
-
-class MemoryTool(ContextPipelineTool):
-    
-    def __init__(self,configService:ConfigService,graphitiService:GraphitiService):
-        super().__init__()
-        self.configService = configService
-        self.graphitiService = graphitiService
-    
-    async def __call__(self,query:str ):
-        ...
